Fixing notifications on 31+

+++ app/src/main/java/com/advice/schedule/database/ReminderManager.kt
@@ -19,7 +19,7 @@ class ReminderManager(
         private const val TAG = "reminder_"
     }
 
-    suspend fun getEvent(conference: String, id: Int): Event? {
+    suspend fun getEvent(conference: String, id: Long): Event? {
         return databaseManager.getEventById(conference, id)
     }
